Removed our own test and moved them to its own file
